+++ src/Actions/AuthAction.js
@@ -2,10 +2,10 @@ import authService from "../Services/AuthService";
 
 export default class AuthActions {
 
-  static async login(data) {
-    return authService.login(data).then(res => {
-      return res.data || {};
-    });
-  }
+  // static async login(data) {
+  //   return authService.login(data).then(res => {
+  //     return res.data || {};
+  //   });
+  // }
 
 }
