+++ src/Actions/CarAction.js
@@ -3,9 +3,9 @@ import CarService from "../Services/CarService";
 export default class CarActions {
 
   static async getCars() {
-    return CarService.getCars().then(res => {
-      return res.data || {};
-    });
+    // return CarService.getCars().then(res => {
+    //   return res.data || {};
+    // });
   }
 
 }
