+++ src/Common/RestAPiInstance.js
@@ -4,9 +4,10 @@ let defautRequestConfig = {
   baseURL: '/'
 }
 
-const defaultRestApi = axios.create(defautRequestConfig);
+// const defaultRestApi = axios.create(defautRequestConfig);
+const defaultRestApi = ''
 
-defaultRestApi.interceptors.request.use(function (config) {
+/*defaultRestApi.interceptors.request.use(function (config) {
     // config.params = {
     //   locale: "en_US",
     //   '_': new Date().getTime(),
@@ -16,6 +17,6 @@ defaultRestApi.interceptors.request.use(function (config) {
   }, function (error) {
     // Do something with request error
     return Promise.reject(error);
-  });
+  });*/
 
 export default defaultRestApi;
