+++ src/Components/Customer/AddCustomer.jsx
@@ -2,19 +2,17 @@ import * as React from "react";
 import Button from "@mui/material/Button";
 import CssBaseline from "@mui/material/CssBaseline";
 import TextField from "@mui/material/TextField";
-import Link from "@mui/material/Link";
 import Grid from "@mui/material/Grid";
-import { Box, Avatar, Typography } from "@mui/material";
+import Box from "@mui/material/Box";
 import Container from "@mui/material/Container";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { useFormik } from "formik";
 import { registerCustomer } from "../../Actions/UserAction";
 import { useNavigate } from "react-router-dom";
-import CarRentalIcon from "@mui/icons-material/CarRental";
 
 const defaultTheme = createTheme();
 
-export default function SignUp() {
+export default function AddCustomer() {
   const navigate = useNavigate();
 
   const formik = useFormik({
@@ -50,41 +48,6 @@ export default function SignUp() {
     <ThemeProvider theme={defaultTheme}>
       <Container component="main" maxWidth="xs">
         <CssBaseline />
-        <Box sx={{ flexGrow: 1 }}>
-          <Grid container spacing={2}>
-            <Grid item xs={12}>
-              <Box
-                sx={{
-                  my: 2,
-                  mx: 2,
-                  display: "flex",
-                  flexDirection: "column",
-                  alignItems: "center",
-                }}
-              >
-                <Typography
-                  style={{ color: "#2196f3" }}
-                  component="h4"
-                  variant="h4"
-                >
-                  Signup for Car Rentals
-                </Typography>
-
-                <Avatar sx={{ m: 1, bgcolor: "secondary.main" }}>
-                  <CarRentalIcon />
-                </Avatar>
-
-                <Typography
-                  style={{ color: "#2196f3" }}
-                  component="h6"
-                  variant="h6"
-                >
-                  Register
-                </Typography>
-              </Box>
-            </Grid>
-          </Grid>
-        </Box>
         <Box
           sx={{
             marginTop: 4,
@@ -194,19 +157,8 @@ export default function SignUp() {
               variant="outlined"
               sx={{ mt: 3, mb: 2 }}
             >
-              Register
+              Add Customer
             </Button>
-            <Grid container justifyContent="flex-end">
-              <Grid item>
-                <Link href="#" variant="body2">
-                  Already have an account?
-                  <Box component={"span"}>
-                    {" "}
-                    <Button onClick={() => navigate("/login")}> LogIn</Button>
-                  </Box>
-                </Link>
-              </Grid>
-            </Grid>
           </Box>
         </Box>
       </Container>
