--- src/Components/Navigation/MiniDrawer.jsx
@@ -1,248 +1,0 @@
-import * as React from "react";
-import { styled, useTheme } from "@mui/material/styles";
-import Box from "@mui/material/Box";
-import MuiDrawer from "@mui/material/Drawer";
-import MuiAppBar from "@mui/material/AppBar";
-import Toolbar from "@mui/material/Toolbar";
-import List from "@mui/material/List";
-import CssBaseline from "@mui/material/CssBaseline";
-import Typography from "@mui/material/Typography";
-import Divider from "@mui/material/Divider";
-import IconButton from "@mui/material/IconButton";
-import MenuIcon from "@mui/icons-material/Menu";
-import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import ChevronRightIcon from "@mui/icons-material/ChevronRight";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-import Button from "@mui/material/Button";
-import { useLocation } from "react-router-dom";
-import Reservation from "../Customer/Reservation";
-import { RentalHistory } from "../ResuableComponents/RentalHistory";
-import PaymentDetails from "../Payment/PaymentDetails";
-import DashboardComponent from "../ResuableComponents/DashboardComponent";
-import { CarRegister } from "../Car/CarRegister";
-import { useNavigate } from "react-router-dom";
-
-const drawerWidth = 240;
-
-const openedMixin = (theme) => ({
-  width: drawerWidth,
-  transition: theme.transitions.create("width", {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.enteringScreen,
-  }),
-  overflowX: "hidden",
-});
-
-const closedMixin = (theme) => ({
-  transition: theme.transitions.create("width", {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  overflowX: "hidden",
-  width: `calc(${theme.spacing(7)} + 1px)`,
-  [theme.breakpoints.up("sm")]: {
-    width: `calc(${theme.spacing(8)} + 1px)`,
-  },
-});
-
-const DrawerHeader = styled("div")(({ theme }) => ({
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "flex-end",
-  padding: theme.spacing(0, 1),
-  // necessary for content to be below app bar
-  ...theme.mixins.toolbar,
-}));
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  zIndex: theme.zIndex.drawer + 1,
-  transition: theme.transitions.create(["width", "margin"], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(["width", "margin"], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
-
-const Drawer = styled(MuiDrawer, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  width: drawerWidth,
-  flexShrink: 0,
-  whiteSpace: "nowrap",
-  boxSizing: "border-box",
-  ...(open && {
-    ...openedMixin(theme),
-    "& .MuiDrawer-paper": openedMixin(theme),
-  }),
-  ...(!open && {
-    ...closedMixin(theme),
-    "& .MuiDrawer-paper": closedMixin(theme),
-  }),
-}));
-
-export default function MiniDrawer({
-  title = "Your Dashboard",
-  icons,
-  userContainers,
-}) {
-  const theme = useTheme();
-  const [open, setOpen] = React.useState(false);
-  const [selectedContainerIndex, setSelectedContainerIndex] = React.useState(0);
-
-  const [customerActivities, setCustomerActivities] = React.useState(null);
-  const navigate = useNavigate();
-
-  let { pathname } = useLocation();
-
-  React.useEffect(() => {
-    if (pathname === "/customer") {
-      if (selectedContainerIndex === 1) {
-        navigate('/reservations')
-      } else if (selectedContainerIndex === 2) {
-
-        navigate('/rentalhistory')
-      } else if (selectedContainerIndex === 3)
-
-        navigate('/paymentdetails')
-    } else if (selectedContainerIndex === 0) {
-      setCustomerActivities(null);
-    } else if (pathname === "/admin" || pathname === "/manager") {
-      if (selectedContainerIndex === 3) {
-
-        navigate('/car-register')
-      } else if (selectedContainerIndex === 2) {
-
-        navigate('/car-register')
-      }
-    }
-
-  }, [selectedContainerIndex]);
-
-  const handleDrawerOpen = () => {
-    setOpen(true);
-  };
-
-  const handleDrawerClose = () => {
-    setOpen(false);
-  };
-
-  const handleSelectedContainer = (index) => {
-    setSelectedContainerIndex(index);
-  };
-
-  // debugger;
-  // console.log(pathname);
-
-  return (
-    <Box sx={{ display: "flex" }}>
-      <CssBaseline />
-      {/* <AppBar position="fixed" open={open}> */}
-      <AppBar position="fixed" open={open}>
-        <Toolbar>
-          <IconButton
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            edge="start"
-            sx={{
-              marginRight: 5,
-              ...(open && { display: "none" }),
-            }}
-          >
-            <MenuIcon />
-          </IconButton>
-
-          <Typography variant="h6" component="h5">
-            Company's logo
-          </Typography>
-          <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-            {title}
-          </Typography>
-          <Button color="inherit">Logout</Button>
-        </Toolbar>
-      </AppBar>
-      <Drawer variant="permanent" open={open}>
-        <DrawerHeader>
-          <IconButton onClick={handleDrawerClose}>
-            {theme.direction === "rtl" ? (
-              <ChevronRightIcon />
-            ) : (
-              <ChevronLeftIcon />
-            )}
-          </IconButton>
-        </DrawerHeader>
-
-        <List>
-          {icons.map((item, index) => (
-            <>
-              <Divider />
-              <ListItem
-                key={item.title}
-                disablePadding
-                sx={{ display: "block" }}
-              >
-                {/* {pathname === "/customer" ? setCustomerSelection(index) : -1} */}
-                <ListItemButton
-                  onClick={() => handleSelectedContainer(index)}
-                  sx={{
-                    minHeight: 48,
-                    justifyContent: open ? "initial" : "center",
-                    px: 2.5,
-                  }}
-                >
-                  <ListItemIcon
-                    sx={{
-                      minWidth: 0,
-                      mr: open ? 3 : "auto",
-                      justifyContent: "center",
-                    }}
-                  >
-                    {item.icon}
-                  </ListItemIcon>
-                  <ListItemText
-                    primary={item.title}
-                    sx={{ opacity: open ? 1 : 0 }}
-                  />
-                </ListItemButton>
-              </ListItem>
-            </>
-          ))}
-        </List>
-        <Divider />
-      </Drawer>
-      <Box
-        component="main"
-        sx={{
-          flexGrow: 1,
-          paddingTop: 8,
-          paddingRight: 2,
-          paddingLeft: 2,
-          paddingBottom: 2,
-          // border: "2px blue groove",
-        }}
-      >
-        {customerActivities ? (
-          customerActivities
-        ) : (
-          <DashboardComponent
-            userContainers={userContainers}
-            // setSelectedContainerIndex={setSelectedContainerIndex}
-            selectedContainerIndex={selectedContainerIndex}
-          />
-        )}
-      </Box>
-    </Box>
-  );
-}
