--- src/Components/TestComponents/AddCustomer.jsx
@@ -1,174 +1,0 @@
-import * as React from "react";
-import Button from "@mui/material/Button";
-import CssBaseline from "@mui/material/CssBaseline";
-import TextField from "@mui/material/TextField";
-import Link from "@mui/material/Link";
-import Grid from "@mui/material/Grid";
-import Box from "@mui/material/Box";
-import Container from "@mui/material/Container";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
-import { useFormik } from "formik";
-import { registerCustomer } from "../../Actions/UserAction";
-import { useNavigate } from "react-router-dom";
-
-const defaultTheme = createTheme();
-
-export default function AddCustomer() {
-
-    const navigate = useNavigate();
-
-    const formik = useFormik({
-        initialValues: {
-            firstName: "",
-            lastName: "",
-            username: "",
-            phone: "",
-            email: "",
-            password: "",
-            // confirmpassword: "",
-        },
-        onSubmit: (values, { resetForm }) => {
-            console.log(values);
-            // navigate("/admin");
-            registerCustomer(values).then(res => {
-                console.log(res)
-                resetForm();
-                navigate("/home/customer")
-            }).catch(err => {
-                // resetForm();
-                console.log(err)
-                resetForm()
-            })
-        },
-    });
-
-    // console.log("adf");
-
-    return (
-        <ThemeProvider theme={defaultTheme}>
-            <Container component="main" maxWidth="xs">
-                <CssBaseline />
-                <Box
-                    sx={{
-                        marginTop: 4,
-                        display: "flex",
-                        flexDirection: "column",
-                        alignItems: "center",
-                    }}
-                >
-                    <Box
-                        component="form"
-                        noValidate
-                        onSubmit={formik.handleSubmit}
-                        sx={{ mt: 3 }}
-                    >
-                        <Grid container spacing={2}>
-                            <Grid item xs={12} sm={6}>
-                                <TextField
-                                    autoComplete="given-name"
-                                    name="firstName"
-                                    required
-                                    fullWidth
-                                    id="firstName"
-                                    label="First Name"
-                                    autoFocus
-                                    value={formik.values.firstName}
-                                    onChange={formik.handleChange}
-                                />
-                            </Grid>
-                            <Grid item xs={12} sm={6}>
-                                <TextField
-                                    required
-                                    fullWidth
-                                    id="lastName"
-                                    label="Last Name"
-                                    name="lastName"
-                                    autoComplete="family-name"
-                                    value={formik.values.lastName}
-                                    onChange={formik.handleChange}
-                                />
-                            </Grid>
-                            <Grid item xs={12} sm={6}>
-                                <TextField
-                                    required
-                                    fullWidth
-                                    id="username"
-                                    label="Username"
-                                    name="username"
-                                    autoComplete="family-name"
-                                    value={formik.values.username}
-                                    onChange={formik.handleChange}
-                                />
-                            </Grid>
-                            <Grid item xs={12} sm={6}>
-                                <TextField
-                                    required
-                                    fullWidth
-                                    id="phone"
-                                    label="Phone"
-                                    name="phone"
-                                    autoComplete="family-name"
-                                    value={formik.values.phone}
-                                    onChange={formik.handleChange}
-                                />
-                            </Grid>
-                            <Grid item xs={12}>
-                                <TextField
-                                    required
-                                    fullWidth
-                                    id="email"
-                                    label="Email Address"
-                                    name="email"
-                                    autoComplete="email"
-                                    value={formik.values.email}
-                                    onChange={formik.handleChange}
-                                />
-                            </Grid>
-                            <Grid item xs={12}>
-                                <TextField
-                                    required
-                                    fullWidth
-                                    name="password"
-                                    label="Password"
-                                    type="password"
-                                    id="password"
-                                    autoComplete="new-password"
-                                    value={formik.values.password}
-                                    onChange={formik.handleChange}
-                                />
-                            </Grid>
-                            {/* <Grid item xs={12}>
-                <TextField
-                  required
-                  fullWidth
-                  name="confirmpassword"
-                  label="Confirm Password"
-                  type="confirmpassword"
-                  id="confirmpassword"
-                  autoComplete="new-password"
-                  value={formik.values.confirmpassword}
-                  onChange={formik.handleChange}
-                />
-              </Grid> */}
-                        </Grid>
-                        <Button
-                            type="submit"
-                            fullWidth
-                            variant="outlined"
-                            sx={{ mt: 3, mb: 2 }}
-                        >
-                            Add Customer
-                        </Button>
-                        {/* <Grid container justifyContent="flex-end">
-                            <Grid item>
-                                <Link href="#" variant="body2">
-                                    Already have an account? Sign In
-                                </Link>
-                            </Grid>
-                        </Grid> */}
-                    </Box>
-                </Box>
-            </Container>
-        </ThemeProvider>
-    );
-}
