--- src/Components/TestComponents/AuthContext.jsx
@@ -1,16 +1,0 @@
-import { createContext } from 'react';
-import React, { useState } from 'react';
-
-const AuthContext = createContext();
-
-export const AuthProvider = ({ children }) => {
-    const [authToken, setAuthToken] = useState(null);
-
-    return (
-        <AuthContext.Provider value={{ authToken, setAuthToken }}>
-            {children}
-        </AuthContext.Provider>
-    );
-};
-
-export default AuthContext;
