--- src/Components/TestComponents/RootLayout.jsx
@@ -1,14 +1,0 @@
-import { Box } from "@mui/material";
-import { Outlet } from "react-router-dom";
-import Header from "./Header";
-
-export default function RootLayout() {
-    return (
-        <Box>
-            {/* <Header /> */}
-            <main>
-                <Outlet />
-            </main>
-        </Box>
-    );
-}
