+++ src/Components/Welcome/WelcomePage.jsx
@@ -1,0 +1,9 @@
+import { Typography } from "@mui/material";
+
+export const WelcomePage = () => {
+  return (
+    <Typography variant="h2" align="center">
+      Welcome to car fleet system
+    </Typography>
+  );
+};
