+++ src/Services/AuthService.js
@@ -4,12 +4,12 @@ import APIs from '../Common/API';
 
 export default class AuthService {
 
-  static login = (data) => {
-    return defaultRestApi.post(APIs.getToken(), data);
-  }
+  // static login = (data) => {
+  //   return defaultRestApi.post(APIs.getToken(), data);
+  // }
 
-  static registerCustomer = (data) => {
-    return defaultRestApi.post(APIs.register(), data);
-  }
+  // static registerCustomer = (data) => {
+  //   return defaultRestApi.post(APIs.register(), data);
+  // }
 
 }
